+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Technical: updated all backend dependencies
 - Technical: updated the frontend dependencies
 - API: added a new "short" mode for the observations (table) API endpoint
+- moved some endpoints from the internal to the public API (to reflect external usage)
 
 # v1.7.3 (2024-03-26)